@@ -4,16 +4,16 @@ abstract class RaeEvent extends Equatable {
   const RaeEvent();
 
   @override
-  List<Object> get props => [];
+  List<Object?> get props => [];
 }
 
 class RaeSubmit extends RaeEvent {
-  final String word;
+  final String? word;
 
   RaeSubmit({this.word});
 
   @override
-  List<Object> get props => [word];
+  List<Object?> get props => [word];
 }
 
 
