@@ -3,11 +3,11 @@ import 'package:rae_test/extension/context_extension.dart';
 
 class NotFoundAlertWidget extends StatelessWidget {
   const NotFoundAlertWidget({
-    Key key,
-    @required String word,
+    Key? key,
+    required String? word,
   }) : _word = word, super(key: key);
 
-  final String _word;
+  final String? _word;
 
   @override
   Widget build(BuildContext context) {
@@ -23,7 +23,7 @@ class NotFoundAlertWidget extends StatelessWidget {
           padding:
               EdgeInsets.only(left: context.pcw(5)),
           child: Text(
-            'La Rae no recoge la palabra \n"${_word.toUpperCase()}"',
+            'La Rae no recoge la palabra \n"${_word!.toUpperCase()}"',
             maxLines: 2,
             style: TextStyle(
                 fontSize: context.pcw(4),
