@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 class SearchFABWidget extends StatelessWidget {
   const SearchFABWidget({
-    Key key,
-    @required GlobalKey<FormState> formKey,
+    Key? key,
+    required GlobalKey<FormState> formKey,
   })  : _formKey = formKey,
         super(key: key);
 
@@ -13,8 +13,8 @@ class SearchFABWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return FloatingActionButton(
       onPressed: () async {
-        if (_formKey.currentState.validate()) {
-          _formKey.currentState.save();
+        if (_formKey.currentState!.validate()) {
+          _formKey.currentState!.save();
         }
       },
       tooltip: 'Consultar',
