@@ -3,7 +3,7 @@ import 'package:lumberdash/lumberdash.dart';
 
 class MyBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object event) {
+  void onEvent(Bloc bloc, Object? event) {
     super.onEvent(bloc, event);
     logMessage('BlocObserver :: Event --> $event');
   }
